+++ src/main/java/electronicDiary_35235/services/JsonReader.java
@@ -1,7 +1,8 @@
-package electronicDiary_35235;
+package electronicDiary_35235.services;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import electronicDiary_35235.entity.Student;
 import lombok.NonNull;
 
 import java.io.*;
@@ -10,7 +11,7 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 
 public class JsonReader {
-    public List<Student> readJson(@NonNull String fileName) {
+    public List<Student> readJsonFromFile(@NonNull String fileName) {
         ObjectMapper objectMapper = new ObjectMapper();
         try(BufferedReader br= new BufferedReader (new InputStreamReader( new FileInputStream(fileName), StandardCharsets.UTF_8))) {
             return objectMapper.readValue(br, new TypeReference<>() {
